pyproject.toml
[project]
name = "lumen2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "matlabengine",
    "numpy>=2.4.2",
    "scipy>=1.17.1",
    "scipy-stubs~=1.17.1",
]

[project.optional-dependencies]
dev = ["ruff"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ARG"]
ignore = ["E203", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
docstring-code-format = true
